Draw loading and empty chart states in MetricCard
Loading now shows the same redacted placeholder sparkline the release
rows use, and the empty state keeps the chart's gridlines instead of
collapsing to blank space. The macOS system gray stand-ins were
duplicated in two files, so they move to a shared Platform file.
