+++ Sources/Scout/UI/Insights/Chart/View/MiniChart.swift
@@ -57,21 +57,6 @@ struct MiniChart: View {
     }
 }
 
-#if os(macOS)
-    import AppKit
-
-    extension NSColor {
-        // Stand-in for the iOS-only tiered system gray, matching its light and dark values.
-        fileprivate static var systemGray3: NSColor {
-            NSColor(name: nil) { appearance in
-                appearance.bestMatch(from: [.darkAqua, .aqua]) == .darkAqua
-                    ? NSColor(red: 72 / 255, green: 72 / 255, blue: 74 / 255, alpha: 1)
-                    : NSColor(red: 199 / 255, green: 199 / 255, blue: 204 / 255, alpha: 1)
-            }
-        }
-    }
-#endif
-
 #Preview {
     NavigationStack {
         List {
