+++ Sources/Scout/UI/Primitives/Cards/MetricCard.swift
@@ -34,21 +34,27 @@ struct MetricCard: View {
 
             if let series = summary?.series {
                 Sparkline(series: series, color: color)
+            } else if summary != nil {
+                Sparkline(series: .placeholder, color: Color(.systemGray3)).redacted(reason: .placeholder)
             } else {
-                Color.clear
+                Sparkline(series: .empty, color: color)
             }
         }
     }
 }
 
-#Preview {
-    HStack(spacing: 24) {
-        MetricCard(
-            summary: MetricSummary(count: 8420, previous: 7500, values: [3, 5, 4, 7, 6, 9, 12]),
-            color: .purple
-        )
+#Preview("MetricCard") {
+    VStack(spacing: 24) {
+        let summary = MetricSummary(count: 8420, previous: 7500, values: [3, 5, 4, 7, 6, 9, 12])
+
+        MetricCard(summary: summary, color: .purple)
+            .frame(width: 162, height: 100)
+
         MetricCard(summary: .loading, color: .green)
+            .frame(width: 162, height: 100)
+
         MetricCard(summary: nil, color: .green)
+            .frame(width: 162, height: 100)
     }
     .padding()
 }
